+++ src/index.ts
@@ -1,3 +1,4 @@
 export { AuthenticationClient } from './auth';
-export { OSSClient } from './oss';
-export { ModelDerivativeClient, urnify } from './md';
+export { OSSClient, DataRetentionPolicy } from './oss';
+export { DataManagementClient } from './dm';
+export { ModelDerivativeClient, ThumbnailSize, urnify } from './md';