+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.6] - 2021-08-20
+
+### Added
+
+- Initial support for browsing DM hubs, projects, folders, items, and versions
+- Support for querying profile of authenticated user
+
 ## [0.0.5] - 2021-08-20
 
 ### Added